Renamed MPITest2 to MPITest. MPITest now tests batches
