+++ PVSystemTests/MPITest/input/MPI_test.params
@@ -15,15 +15,16 @@ HyPerCol "column" = {
    dt = 1.0;
    randomSeed = 41896532;  // if not set here,  clock time is used to generate seed
    startTime = 0.0;
-   stopTime = 100.0;
+   stopTime = 20.0;
    progressStep = 200;
    writeProgressToErr = false;
    outputPath = "output/";
-   filenamesContainLayerNames = false;
-   filenamesContainConnectionNames = false;
+   filenamesContainLayerNames = true;
+   filenamesContainConnectionNames = true;
    checkpointRead = false;
    checkpointWrite = false;
    suppressLastOutput = false;
+   nbatch = 2;
 };
 
 //
