+++ PVSystemTests/MPITest/linux-cmake.cproject
@@ -33,7 +33,7 @@
 		</cconfiguration>
 	</storageModule>
 	<storageModule moduleId="cdtBuildSystem" version="4.0.0">
-		<project id="MPITest2.null.2138707182" name="MPITest2"/>
+		<project id="MPITest.null.2138707182" name="MPITest"/>
 	</storageModule>
 	<storageModule moduleId="scannerConfiguration">
 		<autodiscovery enabled="true" problemReportingEnabled="true" selectedProfileId=""/>
