+++ PVSystemTests/MPITest/mac-cmake.cproject
@@ -35,7 +35,7 @@
 		</cconfiguration>
 	</storageModule>
 	<storageModule moduleId="cdtBuildSystem" version="4.0.0">
-		<project id="MPITest2.null.880059812" name="MPITest2"/>
+		<project id="MPITest.null.880059812" name="MPITest"/>
 	</storageModule>
 	<storageModule moduleId="scannerConfiguration">
 		<autodiscovery enabled="true" problemReportingEnabled="true" selectedProfileId=""/>
